# support/compressxdr: Fix compressxdr performance