--- a/support/compressxdr/compress_xdr_test.go
+++ b/support/compressxdr/compress_xdr_test.go
@@ -2,10 +2,13 @@ package compressxdr
 
 import (
 	"bytes"
+	"io/ioutil"
+	"os"
 	"testing"
 
-	"github.com/stellar/go/xdr"
 	"github.com/stretchr/testify/require"
+
+	"github.com/stellar/go/xdr"
 )
 
 func createTestLedgerCloseMetaBatch(startSeq, endSeq uint32, count int) xdr.LedgerCloseMetaBatch {
@@ -44,5 +47,36 @@ func TestEncodeDecodeLedgerCloseMetaBatch(t *testing.T) {
 	require.Equal(t, len(testData.LedgerCloseMetas), len(decodedData.LedgerCloseMetas))
 	for i := range testData.LedgerCloseMetas {
 		require.Equal(t, testData.LedgerCloseMetas[i], decodedData.LedgerCloseMetas[i])
+	}
+}
+
+func BenchmarkDecodeLedgerCloseMetaBatch(b *testing.B) {
+	lcmBatch := xdr.LedgerCloseMetaBatch{}
+	decoder := NewXDRDecoder(DefaultCompressor, &lcmBatch)
+
+	for n := 0; n < b.N; n++ {
+		file, err := os.Open("testdata/FCD285FF--53312000.xdr.zstd")
+		require.NoError(b, err)
+
+		_, err = decoder.ReadFrom(file)
+		require.NoError(b, err)
+	}
+}
+
+func BenchmarkEncodeLedgerCloseMetaBatch(b *testing.B) {
+	lcmBatch := xdr.LedgerCloseMetaBatch{}
+	decoder := NewXDRDecoder(DefaultCompressor, &lcmBatch)
+	file, err := os.Open("testdata/FCD285FF--53312000.xdr.zstd")
+	require.NoError(b, err)
+
+	_, err = decoder.ReadFrom(file)
+	require.NoError(b, err)
+
+	b.ResetTimer()
+
+	for n := 0; n < b.N; n++ {
+		encoder := NewXDREncoder(DefaultCompressor, &lcmBatch)
+		_, err = encoder.WriteTo(ioutil.Discard)
+		require.NoError(b, err)
 	}
 }